% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2.R
\name{infoBox2}
\alias{infoBox2}
\title{Create an info box for the main body of a dashboard.}
\usage{
infoBox2(title, value = NULL, subtitle = NULL,
  icon = shiny::icon("bar-chart"), color = "aqua", width = 4,
  href = NULL, fill = FALSE, prog = NULL, prog_width = "70\%")
}
\arguments{
\item{title}{Title text.}

\item{value}{The value to display in the box. Usually a number or short text.}

\item{subtitle}{Subtitle text (optional).}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{color}{A color for the box. Valid colors are listed in
\link{validColors}.}

\item{href}{An optional URL to link to.}

\item{fill}{If \code{FALSE} (the default), use a white background for the
content, and the \code{color} argument for the background of the icon. If
\code{TRUE}, use the \code{color} argument for the background of the
content; the icon will use the same color with a slightly darkened
background.}

\item{prog}{Process text (optional).}

\item{prog_width}{The value of process.}
}
\description{
An info box displays a large icon on the left side, and a title, value
(usually a number), and an optional smaller subtitle on the right side. Info
boxes are meant to be placed in the main body of a dashboard.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(shiny)
library(shinydashboard)

server <- function(input, output) {
  output$infoBox_process <- renderInfoBox({
    infoBox2(
      'Info', 'some text', width = 6,
      fill = TRUE, prog = "Process", prog_width = '70\%'
    )
  })

}

body <- dashboardBody(
  infoBoxOutput(outputId = 'infoBox_process')
)

ui <- dashboardPage(
  dashboardHeader(disable = TRUE),
  dashboardSidebar(disable = TRUE),
  body
)

shinyApp(ui, server)
}
}
\seealso{
Other boxes: \code{\link{box2}}
}
