% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numRangeInput.R
\name{numRangeInput}
\alias{numRangeInput}
\title{Create numeric range input}
\usage{
numRangeInput(inputId, label = NULL, start = "", end = "", min = NULL,
  max = NULL, format = "#,###", separator = "to", width = NULL)
}
\description{
Create an input control for numeric range values
}
\examples{

if (interactive()) {
library(shiny)
library(shinydashboard)
shinyApp(
  ui = fluidPage({
    title = "Num Range Input"
    fluidRow(
      numRangeInput("num", label = "数值：", start = 12, end = 20000),
      sliderInput("obs", "Number of observations:", min = 0, max = 2000, value = c(13,1900)),
      actionButton("update", label = "Update"),
      verbatimTextOutput("value1"),
      verbatimTextOutput("value2")
    )
  }),
  server = function(input, output, session) {
    output$value1 <- renderText({
      paste(input$num_1,input$num_2)
    })
    observeEvent(input$update, {
      updatenumRangeInput(session, "num", range = 1, value = list(input$obs[[1]]))
      updatenumRangeInput(session, "num", range = 2,value = list(input$obs[[2]]))
    })
    output$value2 <- renderText({
      input$obs
    })
  }
)
}
}
